downgrade torchvision version from 0.7.0 to 0.6.0 for offical pytorch 1.5.0

+++ examples/pytorch/vision/torchvision/csrc/cpu/video_reader/VideoReader.cpp
@@ -311,7 +311,7 @@ torch::List<torch::Tensor> readVideo(
         videoFrame = torch::zeros(
             {numVideoFrames, outHeight, outWidth, numChannels}, torch::kByte);
         expectedWrittenBytes =
-            (size_t)numVideoFrames * outHeight * outWidth * numChannels;
+            numVideoFrames * outHeight * outWidth * numChannels;
       }
 
       videoFramePts = torch::zeros({numVideoFrames}, torch::kLong);
